get the rest of my rarely used junk out of common/common
