+++ ui/ceval/src/ctrl.ts
@@ -6,7 +6,8 @@ import { defaultPosition, setupPosition } from 'chessops/variant';
 import { parseFen } from 'chessops/fen';
 import { lichessRules } from 'chessops/compat';
 import { povChances } from './winningChances';
-import { prop, Prop, Toggle, toggle, clamp } from 'common';
+import { prop, Prop, Toggle, toggle } from 'common';
+import { clamp } from 'common/algo';
 import { Result } from '@badrap/result';
 import { storedIntProp, storage, tempStorage } from 'common/storage';
 import { Rules } from 'chessops';
