+++ ui/common/src/common.ts
@@ -110,58 +110,6 @@ export const escapeHtml = (str: string): string =>
         .replace(/"/g, '&quot;')
     : str;
 
-export const clamp = (value: number, bounds: { min?: number; max?: number }): number =>
-  Math.max(bounds.min ?? -Infinity, Math.min(value, bounds.max ?? Infinity));
-
-export function shuffle<T>(array: T[]): T[] {
-  for (let i = array.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [array[i], array[j]] = [array[j], array[i]];
-  }
-  return array;
-}
-
-export function deepFreeze<T>(obj: T): T {
-  if (obj !== null && typeof obj === 'object')
-    Object.values(obj)
-      .filter(v => v !== null && typeof v === 'object')
-      .forEach(o => deepFreeze(o));
-  return Object.freeze(obj);
-}
-
-// recursive comparison of enumerable primitives. complex properties get reference equality only
-export function isEquivalent(a: any, b: any): boolean {
-  if (a === b) return true;
-  if (typeof a !== typeof b) return false;
-  if (Array.isArray(a)) {
-    return Array.isArray(b) && a.length === b.length && a.every((x, i) => isEquivalent(x, b[i]));
-  }
-  if (typeof a !== 'object') return false;
-  const [aKeys, bKeys] = [Object.keys(a), Object.keys(b)];
-  if (aKeys.length !== bKeys.length) return false;
-  for (const key of aKeys) {
-    if (!bKeys.includes(key) || !isEquivalent(a[key], b[key])) return false;
-  }
-  return true;
-}
-
-export function zip<T, U>(arr1: T[], arr2: U[]): [T, U][] {
-  const length = Math.min(arr1.length, arr2.length);
-  const result: [T, U][] = [];
-  for (let i = 0; i < length; i++) {
-    result.push([arr1[i], arr2[i]]);
-  }
-  return result;
-}
-
-export function findMapped<T, U>(arr: T[], callback: (el: T) => U | undefined): U | undefined {
-  for (const el of arr) {
-    const result = callback(el);
-    if (result) return result;
-  }
-  return undefined;
-}
-
 export function frag<T extends Node = Node>(html: string): T {
   const div = document.createElement('div');
   div.innerHTML = html;
